Time: 9 ms (8.24%), Space: 54.6 MB (54.70%) - LeetHub

+++ 2366-minimum-replacements-to-sort-the-array/2366-minimum-replacements-to-sort-the-array.java
@@ -1,7 +1,7 @@
 class Solution {
     public long minimumReplacement(int[] nums) {
         long ans = 0;
-        long value = nums[nums.length - 1];
+        double value = nums[nums.length - 1];
 
         for (int i = nums.length - 2; 0 <= i; i--) {
             if (nums[i] < value) {
@@ -9,9 +9,9 @@ class Solution {
                 continue;
             }
 
-            long count = (long) Math.ceil((double) nums[i] / (double) value);
+            double count = (int) Math.ceil(nums[i] / value);
             ans += (count - 1);
-            value = (long) Math.floor((double) nums[i] / (double) count);
+            value = (int) Math.floor(nums[i] / count);
         }
 
         return ans;
